moved existing login-features to one file: login.feature
- deleted old features (pupil, teacher, admin)
- moved steps to one file
- deleted old steps (pupil, teacher, admin)
- common steps added
- usage of common steps retightened in other features
