+++ features/homework.feature
@@ -47,7 +47,7 @@ Feature: create different types of homework
         And the teacher has posed a homework
         And teacher logs out
         And student logs in with email <username> and password <password>
-        And the pupil should accept the data protection
+        And student with full age accepts student's data protection
         And the student has submitted that homework
         When the teacher uploads file feedback
         Then both the teacher and student can see and download the feedback
