+++ step_definitions/adminCreateClassWithoutParticipants-steps.js
@@ -14,7 +14,7 @@ Given(/^admin is logged in successfully$/, async function () {
     await  loginPage.performLogin(defaultAdminUsername, defaultAdminPassword)
 });
 Given(/^admin performs the first login steps$/, function () {
-    return firstLogin.firstLoginAdmin();
+    return firstLogin.firstLoginAdminOrTeacher();
 });
 
 When(/^admin creates a class (.*)$/, function (className) {
