+++ step_definitions/login-steps.js
@@ -1,4 +1,3 @@
-'use strict';
 
 const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
 const startPage = require('../page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
@@ -9,17 +8,32 @@ const elementHelpers = require('../runtime/helpers/elementHelpers.js');
 const DashboardPage = require('../page-objects/pages/DashboardPage');
 const common = require('../shared_steps/common-steps.js');
 
-When(/^an admin puts in (.*) and (.*) and click the login-button$/, async function(username,password) {
-	await startPage.clickLoginBtn();
-	await  loginPage.performLogin(username, password);
+/*Then(/^pupil accepts data security, checks the email (.*) and sets password (.*) .$/, function (username, password) {
+	return firstLogin.firstLoginPupilFullAge(username, password);
+});*/
+
+Then(/^a pupil should see the dashboard$/, function () {
+	return loginPage.loginResult();
 });
 
+
+
+/*
+Then(/^the teacher should accept the data protection$/, function() {
+	return firstLogin.firstLoginAdminOrTeacher();
+});*/
+
+Then(/^the teacher-dashboard should have an icon with the teacher's initials$/,function() {
+		return loginPage.loginResult();
+	}
+);
+/*
 Then(
 	/^the user is supposed to accept the data protection agreement$/,
 	function() {
-		return firstLogin.firstLoginAdmin();
+		return firstLogin.firstLoginAdminOrTeacher();
 	}
-);
+);*/
 
 Then(
 	/^the admin-dashboard should have the admin initials$/,
@@ -46,3 +60,17 @@ Then(/^the admin-dashboard should have the following tabs$/, function (LoginText
     return DashboardPage.checkIfTabsAreVisible(LoginTextLabels, shared.loginData.elem.loginTabs);
 	}
 );
+
+
+
+
+Then(/^a user should see a notification$/, function() {
+	return loginPage.wrongLoginResult();
+});
+
+Then(/^the login-page should look like it looked before for (.*)$/, function(
+	username
+) {
+	let filename = 'failed-login-page';
+	return loginData.compareScreenshots(filename);
+});
